Common: Move JSONSerializable to common.types
