+++ monkey/common/types.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from enum import Enum
 from ipaddress import IPv4Address
-from typing import Optional
+from typing import Dict, List, Optional, Union
 from uuid import UUID
 
 from pydantic import ConstrainedInt, PositiveInt
@@ -16,6 +16,16 @@ from common.network.network_utils import address_to_ip_port
 AgentID: TypeAlias = UUID
 HardwareID: TypeAlias = PositiveInt
 MachineID: TypeAlias = PositiveInt
+
+JSONSerializable = Union[  # type: ignore[misc]
+    Dict[str, "JSONSerializable"],  # type: ignore[misc]
+    List["JSONSerializable"],  # type: ignore[misc]
+    int,
+    str,
+    float,
+    bool,
+    None,
+]
 
 
 class NetworkPort(ConstrainedInt):
